requirements.txt
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
pusher>=3.0.0
djangorestframework-simplejwt>=4.7.1